started restructuring
