+++ src/oauth1/mod.rs
@@ -1,0 +1,2 @@
+#[unstable]
+pub mod client;
